Revise insertion of bundles
(see commit 9117e3c56356379e2a6270f278947aaac8411634 from 2021/10/16)
I just realized that, in the original codebase (before I forked it), each call to 'spawn' inserted a bundle into a new entity. In updating the code for Bevy 0.5, I called 'spawn_bundle' and then 'insert_bundle' sequentially. This meant that the two bundles were added to the same entity. Whether or not that affects the semantics, I don't know, but I'm going to be cautious. Perhaps this will be investigated once the program is functional.