Simplify unnecessary for loops
At any given time (in terms of gameplay, not literally), there's one entity with the 'CurrentTetromino' component and one entity with the 'Matrix' component. This enabled me to remove many of the for loops, which made it difficult to reason about the behavior of the code. Note that there are actually multiple entities with the 'CurrentTetromino' component, as 'spawn_current_tetromino' erroneously assigns said component to each block of the spawned tetromino.