Settle on representation of current tetromino
Since there won't be a feature for holding a tetromino (not trying to get sued here; also, it's unoriginal in general), a given tetromino is the current tetromino, so 'CurrentTetromino' adds nothing. Meanwhile, a block that is part of the current tetromino stores a 'MatrixPosition' already, so 'Tetromino' doesn't need a field for that. Finally, 'Tetromino' has no need to store a 'TetrominoType', since it doesn't affect anything except when a new tetromino is spawned.

Note the title of the commit; f4ca1d84b7bda7e859c8f120e425cf9f0c447221 is partially overruled. Each block of the current tetromino is once again given a 'Tetromino' component (I really couldn't figure out an alternative), but unlike in 8bit-pudding's original code, the blocks are immediately stored in a 'Vec' for cohesiveness. (If I could tell the compiler that 'curr_tetromino_query' always gives four entities and collect it into '[(Entity, MatrixPosition); 4]', I would.)

This introduced a lot of compilation errors, since the code had a completely different understanding of the representation of the current tetromino.