Refactor to use aggregate of tetromino positions
Continues refactoring work from previous commit. For now, the positions of the blocks of the tetromino will be stored in a 'Vec'; this may become an array in the future. (The only real difference is that the heap is currently used, with a slight cost in allocation and indirection.)

The errors involving 'move_x' and 'move_y' will be addressed elsewhere; I'll have to see if it's even necessary to use them. In general, I'm not sure how the current rotation system works and how to properly implement rotation.

The errors involving 'Tetromino::SIZES' and 'block' will be addressed elsewhere as well. They stem from the fact that the tetromino type is no longer stored in 'Tetromino', as part of the overarching refactor. I highly doubt that it is useful to do so.