Simplify heap logic
Since the matrix's dimensions are known at the start of a game, using a Vec avoids additional allocations. In addition, a Vec wastes no space compared to a HashSet/BTreeSet and indexing to see if a block is at a certain position is trivially fast. I removed the Heap component, as it doesn't serve a purpose; I'm not sure if that goes against the spirit of ECS.