Use system ordering
I'm not sure that using a unit struct was the best choice for one label, given the numerous derive annotations (which add to the compilation time and executable size). Then again, the size increase is probably insignificant for a game, and since more systems will probably be added in the future, it might be better to start using unit structs as system labels now.