Fix movement system; implement lock delay
When I tried to implement lock delay previously, the gravity would somehow continue even when movement in the y-direction should've been voided as a result of an obstacle. It turns out that the match expression that determined the direction would favor left/rightwards movement over downwards movement, so downwards movement wouldn't be checked if the tetromino couldn't descend further but was moved in the x-direction, which is the point of having a lock delay. This led me to represent and check movement in the x-direction and y-direction separately.