Actually fix wall kicks and reverting position
For a while, the translations for the wall kicks were applied cumulatively, which is blatantly wrong. Very recently (today, in fact), I changed it so that, upon a failed wall kick (i.e., the piece was still colliding with something), the position was reverted to how it was before the original, basic rotation, which is also wrong. This actually fixes the wall kicks (I hope) by applying wall kicks relative to the position after the original, basic rotation and only reverting entirely if all wall kicks fail.