Check for and handle losing case more cleanly
Pieces are spawned at the proper height now (above the top of the board). To accomodate, `can_move` specifically checks if any cell in the heap is occupied: that is, a cell is marked `Some(HeapEntry::Occupied)`. In the `None` case, the given cell is above the board, which is normal due to spawning.

Instead of using a hacky workaround to check for the losing case, the `lock` system checks if any of the blocks of the current piece are above the grid, sending an exit event if so.